% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_user_data.R
\name{log_user_data}
\alias{log_user_data}
\title{Main function to log user data and return browser/navigator/os info}
\usage{
log_user_data(user, application, version, write_db = TRUE)
}
\description{
Use in server function
}
