% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_browser_info.R
\name{set_browser_info}
\alias{set_browser_info}
\title{Set the Shiny input value with browser info}
\usage{
set_browser_info(id = "navigatorInfo", session = getDefaultReactiveDomain())
}
\arguments{
\item{id}{A list with browser info is assigned to input$id}
}
\description{
Sets input$navigatorInfo
}
